requirements.txt
antlr4-python3-runtime==4.8
argon2-cffi==21.1.0
asn1crypto==0.24.0
async-generator==1.10
babel==2.5.1
backcall==0.2.0
bleach==4.1.0
blivet==3.2.2
brlapi==0.6.7
certifi==2021.5.30
cffi==1.11.5
chardet==3.0.4
click==8.0.1
configobj==5.0.6
configparser==5.0.2
configshell-fb==1.1.28
coverage==4.5.1
cryptography==2.3
cupshelpers==1.0
cycler==0.10.0
dataclasses==0.8
dbus-python==1.2.4
decorator==4.2.1
defusedxml==0.7.1
dnf-etckeeper==0.0.0
dnspython==1.15.0
docker-pycreds==0.4.0
entrypoints==0.3
ethtool==0.14
file-magic==0.3.0
gitdb==4.0.7
gitpython==3.1.18
gpg==1.13.1
gssapi==1.5.1
html5lib==0.999999999
humanize==0.5.1
hydra-core==1.1.1
idna==2.5
importlib-metadata==4.8.1
importlib-resources==5.2.2
iniparse==0.4
iotop==0.6
ipaclient==4.8.7
ipalib==4.8.7
ipaplatform==4.8.7
ipapython==4.8.7
ipykernel==5.5.5
ipython-genutils==0.2.0
ipython==7.16.1
ipywidgets==7.6.4
isc==2.0
jedi==0.18.0
jinja2==2.10.1
joblib==1.0.1
jsonpointer==1.10
jsonschema==2.6.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
jwcrypto==0.5.0
kerberos==1.3.0
kitchen==1.2.6
kiwisolver==1.3.1
kmod==0.1
line-profiler==3.3.0
lxml==4.2.3
mako==1.0.6.dev0
markupsafe==0.23
matplotlib==3.3.4
meson==0.55.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netaddr==0.7.19
netifaces==0.10.6
nftables==0.1
nose==1.3.7
notebook==6.4.3
ntplib==0.3.3
numpy==1.19.5
ofed-le-utils==1.0.3
omegaconf==2.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
paste==3.0.8
pathtools==0.1.2
pciutils==2.3.6
pcp==5.0
perf==0.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.2.0
pip==21.1
ply==3.9
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.8.0
psycopg2==2.7.5
ptyprocess==0.7.0
pwquality==1.4.0
pyasn1-modules==0.1.5
pyasn1==0.3.7
pycairo==1.16.3
pycparser==2.14
pycups==1.9.72
pycurl==7.43.0.2
pygments==2.10.0
pygobject==3.28.3
pyinotify==0.9.6
pyopenssl==18.0.0
pyparsing==2.1.10
pyparted==3.11.0
pysocks==1.6.8
python-augeas==0.5.0
python-cpio==0.1
python-dateutil==2.8.2
python-dmidecode==3.12.2
python-ldap==3.1.0
python-linux-procfs==0.6.2
python-meh==0.47.2
python-memcached==1.58
python-yubico==1.3.2
pytz==2017.2
pyudev==0.21.0
pyusb==1.0.0
pyyaml==5.4.1
pyzmq==22.2.1
qrcode==5.1
qtconsole==5.1.1
qtpy==1.11.0
reportlab==3.4.0
requests==2.20.0
rhnlib==2.8.6
rpm==4.14.3
rtslib-fb==2.1.73
schedutils==0.6
scikit-learn==0.24.2
scipy==1.5.4
scons==3.0.1
seaborn==0.11.2
selinux==2.9
send2trash==1.8.0
sentry-sdk==1.3.1
sepolicy==1.1
setools==4.3.0
setproctitle==1.1.10
setuptools==56.0.0
shortuuid==1.0.1
six==1.16.0
sklearn==0.0
slip.dbus==0.6.4
slip==0.6.4
smmap==4.0.0
sos==3.9
sssdconfig==2.3.0
subprocess32==3.5.4
subscription-manager==1.27.16
suds-jurko==0.7.dev0
syspurpose==1.27.16
tempita==0.5.1
termcolor==1.1.0
terminado==0.12.0
testpath==0.5.0
threadpoolctl==2.2.0
torch==1.8.1+cu111
torchaudio==0.8.1
torchtext==0.9.1
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.62.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urlgrabber==4.1.0
urllib3==1.24.2
urwid==1.3.1
virtualenv==15.1.0
wandb==0.12.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yaspin==2.1.0
zipp==3.5.0
